Alterado os tipos dos IDs e Foreign Keys para Integer ao invés de Long

+++ railwise_java/src/main/java/br/edu/unoesc/railwise/controllers/CargaController.java
@@ -18,75 +18,75 @@ import br.edu.unoesc.railwise.repositories.CargaRepository;
 @RequestMapping("/carga")
 public class CargaController {
 
-	@Autowired
-	private CargaRepository cargaRepository;
+    @Autowired
+    private CargaRepository cargaRepository;
 
-	/**
-	 * Retorna as cargas cadastradas na base
-	 * 
-	 * @return lista de todas as cargas cadastradas
-	 */
-	@GetMapping
-	public ResponseEntity<List<Carga>> findAll() {
-		List<Carga> carga = cargaRepository.findAll();
+    /**
+     * Retorna as cargas cadastradas na base
+     * 
+     * @return lista de todas as cargas cadastradas
+     */
+    @GetMapping
+    public ResponseEntity<List<Carga>> findAll() {
+        List<Carga> carga = cargaRepository.findAll();
 
-		return ResponseEntity.ok(carga);
-	}
+        return ResponseEntity.ok(carga);
+    }
 
-	/**
-	 * Busca um registro de carga na base pelo id da carga
-	 * 
-	 * @param id código identificador da carga que se quer buscar
-	 * @return registro da carga com id igual ao informado no parâmetro
-	 */
-	@GetMapping("/{id}")
-	public ResponseEntity<Carga> findById(@PathVariable Long id) {
-		Carga carga = cargaRepository.findById(id).get();
+    /**
+     * Busca um registro de carga na base pelo id da carga
+     * 
+     * @param id código identificador da carga que se quer buscar
+     * @return registro da carga com id igual ao informado no parâmetro
+     */
+    @GetMapping("/{id}")
+    public ResponseEntity<Carga> findById(@PathVariable Integer id) {
+        Carga carga = cargaRepository.findById(id).get();
 
-		return ResponseEntity.ok(carga);
-	}
+        return ResponseEntity.ok(carga);
+    }
 
-	/**
-	 * Salva um novo registro de carga
-	 * 
-	 * @param carga objeto de Carga com as informações da carga a ser cadastrada
-	 * @return o objeto de carga gravado na base
-	 */
-	@PostMapping("/cadastro")
-	public ResponseEntity<Carga> save(@RequestBody Carga carga) {
-		cargaRepository.save(carga);
+    /**
+     * Salva um novo registro de carga
+     * 
+     * @param carga objeto de Carga com as informações da carga a ser cadastrada
+     * @return o objeto de carga gravado na base
+     */
+    @PostMapping("/cadastro")
+    public ResponseEntity<Carga> save(@RequestBody Carga carga) {
+        cargaRepository.save(carga);
 
-		return ResponseEntity.ok(carga);
-	}
+        return ResponseEntity.ok(carga);
+    }
 
-	/**
-	 * Realiza a edição de um registro de carga já cadastrado
-	 * 
-	 * @param id    id da carga que se deseja editar
-	 * @param carga objeto de Carga com as novas informações a serem registradas
-	 * @return objeto de Carga salvo na base
-	 */
-	@PutMapping("/{id}")
-	public ResponseEntity<Carga> update(@PathVariable Long id, @RequestBody Carga carga) {
-		carga.setId_carga(id);
-		cargaRepository.save(carga);
+    /**
+     * Realiza a edição de um registro de carga já cadastrado
+     * 
+     * @param id    id da carga que se deseja editar
+     * @param carga objeto de Carga com as novas informações a serem registradas
+     * @return objeto de Carga salvo na base
+     */
+    @PutMapping("/{id}")
+    public ResponseEntity<Carga> update(@PathVariable Integer id, @RequestBody Carga carga) {
+        carga.setId_carga(id);
+        cargaRepository.save(carga);
 
-		return ResponseEntity.ok(carga);
-	}
+        return ResponseEntity.ok(carga);
+    }
 
-	/**
-	 * Realiza a exclusão de uma carga
-	 * 
-	 * @param id código identificador da carga a ser excluída
-	 * @return retorna um booleano -> true para exclusão bem sucedida
-	 */
-	@DeleteMapping("/{id}")
-	public ResponseEntity<Boolean> delete(@PathVariable Long id) {
-		Carga carga = cargaRepository.findById(id).get();
+    /**
+     * Realiza a exclusão de uma carga
+     * 
+     * @param id código identificador da carga a ser excluída
+     * @return retorna um booleano -> true para exclusão bem sucedida
+     */
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Boolean> delete(@PathVariable Integer id) {
+        Carga carga = cargaRepository.findById(id).get();
 
-		cargaRepository.delete(carga);
+        cargaRepository.delete(carga);
 
-		return ResponseEntity.ok(true);
-	}
+        return ResponseEntity.ok(true);
+    }
 
 }
