+++ railwise_java/src/main/java/br/edu/unoesc/railwise/controllers/ManutencaoController.java
@@ -26,76 +26,76 @@ import br.edu.unoesc.railwise.repositories.ManutencaoRepository;
 @RequestMapping("/manutencao")
 public class ManutencaoController {
 
-	@Autowired
-	private ManutencaoRepository manutencaoRepository;
+    @Autowired
+    private ManutencaoRepository manutencaoRepository;
 
-	/**
-	 * Retorna as manutenções cadastradas na base
-	 * 
-	 * @return lista de todas as manutenções cadastradas
-	 */
-	@GetMapping
-	public ResponseEntity<List<Manutencao>> findAll() {
-		List<Manutencao> manutencao = manutencaoRepository.findAll();
+    /**
+     * Retorna as manutenções cadastradas na base
+     * 
+     * @return lista de todas as manutenções cadastradas
+     */
+    @GetMapping
+    public ResponseEntity<List<Manutencao>> findAll() {
+        List<Manutencao> manutencao = manutencaoRepository.findAll();
 
-		return ResponseEntity.ok(manutencao);
-	}
+        return ResponseEntity.ok(manutencao);
+    }
 
-	/**
-	 * Busca o registro de uma manutenção na base pelo código
-	 * 
-	 * @param id código identificador da manutenção que se quer encontrar
-	 * @return registro da manutenção cujo id corresponde ao id passado no parâmetro
-	 */
-	@GetMapping("/{id}")
-	public ResponseEntity<Manutencao> findById(@PathVariable Long id) {
-		Manutencao manutencao = manutencaoRepository.findById(id).get();
+    /**
+     * Busca o registro de uma manutenção na base pelo código
+     * 
+     * @param id código identificador da manutenção que se quer encontrar
+     * @return registro da manutenção cujo id corresponde ao id passado no parâmetro
+     */
+    @GetMapping("/{id}")
+    public ResponseEntity<Manutencao> findById(@PathVariable Integer id) {
+        Manutencao manutencao = manutencaoRepository.findById(id).get();
 
-		return ResponseEntity.ok(manutencao);
-	}
+        return ResponseEntity.ok(manutencao);
+    }
 
-	/**
-	 * Salva uma nova manutenção na base
-	 * 
-	 * @param manutencao objeto de Manutencao com as informações a serem salvas
-	 * @return o objeto da Manutencao gravada na base
-	 */
-	@PostMapping("/cadastro")
-	public ResponseEntity<Manutencao> save(@RequestBody Manutencao manutencao) {
+    /**
+     * Salva uma nova manutenção na base
+     * 
+     * @param manutencao objeto de Manutencao com as informações a serem salvas
+     * @return o objeto da Manutencao gravada na base
+     */
+    @PostMapping("/cadastro")
+    public ResponseEntity<Manutencao> save(@RequestBody Manutencao manutencao) {
 
-		manutencaoRepository.save(manutencao);
+        manutencaoRepository.save(manutencao);
 
-		return ResponseEntity.ok(manutencao);
-	}
+        return ResponseEntity.ok(manutencao);
+    }
 
-	/**
-	 * Realiza a edição de um registro de manutenção
-	 * 
-	 * @param id         código da manutenção a ser editada
-	 * @param manutencao objeto de Manutencao com as informações a serem registradas
-	 * @return objeto de Manutencao registrado na base
-	 */
-	@PutMapping("/{id}")
-	public ResponseEntity<Manutencao> update(@PathVariable Long id, @RequestBody Manutencao manutencao) {
-		manutencao.setId_manutencao(id);
+    /**
+     * Realiza a edição de um registro de manutenção
+     * 
+     * @param id         código da manutenção a ser editada
+     * @param manutencao objeto de Manutencao com as informações a serem registradas
+     * @return objeto de Manutencao registrado na base
+     */
+    @PutMapping("/{id}")
+    public ResponseEntity<Manutencao> update(@PathVariable Integer id, @RequestBody Manutencao manutencao) {
+        manutencao.setId_manutencao(id);
 
-		manutencaoRepository.save(manutencao);
+        manutencaoRepository.save(manutencao);
 
-		return ResponseEntity.ok(manutencao);
-	}
+        return ResponseEntity.ok(manutencao);
+    }
 
-	/**
-	 * Realiza a exclusão de uma manutenção da base
-	 * 
-	 * @param id código da manutenção a ser excluída
-	 * @return retorna um valor booleano -> true para exclusão bem sucedida
-	 */
-	@DeleteMapping("/{id}")
-	public ResponseEntity<Boolean> delete(@PathVariable Long id) {
-		Manutencao manutencao = manutencaoRepository.findById(id).get();
+    /**
+     * Realiza a exclusão de uma manutenção da base
+     * 
+     * @param id código da manutenção a ser excluída
+     * @return retorna um valor booleano -> true para exclusão bem sucedida
+     */
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Boolean> delete(@PathVariable Integer id) {
+        Manutencao manutencao = manutencaoRepository.findById(id).get();
 
-		manutencaoRepository.delete(manutencao);
+        manutencaoRepository.delete(manutencao);
 
-		return ResponseEntity.ok(true);
-	}
+        return ResponseEntity.ok(true);
+    }
 }
